Bring the normal view more in-line with the original hp48 calculators.
Propose using the 'Modifier Letter Small X' (U+02E3) rather than the 'Superscript Latin Small Letter N' (U+207F) to bring the normal view of Droid48 more in-line with the original characters. If this works it would have the benefit of correctly representing the x position on the RPN stack.

Many thanks for your work on Droid48. I love the app, and noticed that you appear to be preparing an updated version, so I thought I'd propose this now. Thanks again & I hope you're not offended by the suggestion.